Added beat lines to matrix view, and tweaked colours and things
A few other very minor changes.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@1445 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/notationstaff.cpp
@@ -104,7 +104,7 @@ NotationStaff::setLegatoDuration(Rosegarden::timeT duration)
     wq->setUnit(duration);
 }
 
-void NotationStaff::insertTimeSignature(int layoutX,
+void NotationStaff::insertTimeSignature(double layoutX,
 					const TimeSignature &timeSig)
 {
     QCanvasPixmap *pixmap =
